CF-models/COMBINEX/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==1.26.4
omegaconf==2.3.0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pyemd==1.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
sentry-sdk==2.18.0
setproctitle==1.3.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sympy==1.13.1
texttable==1.7.0
threadpoolctl==3.5.0
torcheval==0.0.7
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.17.5
wrapt==1.17.0

# --- PyTorch / PyG kept out of requirements.txt on purpose on Windows ---
# Install them with the commands below (CPU or GPU), so pip doesn't try the wrong wheels.
# python -m pip install --upgrade pip setuptools wheel
# python -m pip install torch==2.5.0 --index-url https://download.pytorch.org/whl/cu124
# python -m pip install pyg-lib torch-scatter torch-sparse torch-cluster torch-spline-conv -f https://data.pyg.org/whl/torch-2.5.0+cu124.html
# python -m pip install torch-geometric==2.6.1